+++ NEWS.md
@@ -1,5 +1,8 @@
 # phylotax (development version)
-
+* **BREAKING CHANGE** Renamed the `phylotax` class elements "`tip_taxa`" ->
+  "`assigned`" and "`node_taxa`" -> "`node_assigned`", to be more consistent
+  with the other class elements (`accepted`, `rejected`, `missing`)
+* Removed dependence on `stringr`.
 * Added function `keep_tips()` to make subsets of the tips in a `phylotax`
   object, including the tree and the taxonomic annotations.
 * Additional utility function `relabel_tree()` maps the tip labels of a tree